Move getFocalPointStatus() method up to ViewController
